tests/wpt/web-platform-tests/shadow-dom/untriaged/elements-and-dom-objects/extensions-to-element-interface/attributes/test-005.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_02_01_05</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#shadow-aware-attributes">
<meta name="assert" content="Extensions to Element Interface: shadowRoot of type ShadowRoot">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function () {

	var d = newHTMLDocument();

    var host = d.createElement('div');
    d.body.appendChild(host);

    //old tree
    var s1 = host.createShadowRoot();
    //young tree
    var s2 = host.createShadowRoot();

    assert_equals(host.shadowRoot, s2, 'attribute shadowRoot must return the youngest tree that has ' +
        	'the context object as its shadow host');


}, 'A_10_02_01_05_T01');
</script>
</body>
</html>
